I noticed that the parsing and formatting operations of KSUIDs were quite inefficient regarding memory allocations, so I spent a bit of time optimizing them.

The biggest gain came from appending to the output buffer when converting from one base to another (instead of prepending) and reversing it at the end. I also changed some internal APIs to make it possible to pass preallocated buffers, and added the `KSUID.Append([]byte) []byte` function to allow formatting of KSUID with zero dynamic memory allocations.

Here are the numbers before the changes:
```
BenchmarkAppend-4   	  200000	      6539 ns/op	    1272 B/op	      95 allocs/op
PASS
ok  	github.com/segmentio/ksuid	1.385s
```
Then after:
```
BenchmarkAppend-4   	  300000	      4297 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/segmentio/ksuid	1.339s
```
And the comparison:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkAppend-4     6539          4297          -34.29%

benchmark             old allocs     new allocs     delta
BenchmarkAppend-4     95             0              -100.00%

benchmark             old bytes     new bytes     delta
BenchmarkAppend-4     1272          0             -100.00%
```

Please take a look and let me know if anything should be changed.