@achille-roussel 2017-06-11
I added another optimization for the `ksuid.NewRandom` method to keep a memory pool of payload buffers so they don't need to be reallocated on the heap for every call.
- Before:
```
$ go test -v -run _ -bench New -benchmem -memprofile /tmp/mem
BenchmarkNew-4   	 1000000	      1243 ns/op	      16 B/op	       1 allocs/op
PASS
ok  	github.com/segmentio/ksuid	1.265s
```
- After:
```
$ go test -v -run _ -bench New -benchmem -memprofile /tmp/mem
BenchmarkNew-4   	 1000000	      1207 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/segmentio/ksuid	1.228s
```